requirements.txt
# Core
numpy
python-dotenv
pypdf

# LLM / embeddings
openai

# Vector index (imported as `faiss`)
faiss-cpu
